+++ @osrs-tracker/models/CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.1.0 - 2023/03/19
+
+- Added `combatLevel: number` to `Player`
+- Combined `Player` and `PlayerWithHiscores`.
+
 ## v0.0.3 - 2023/03/16
 
 - made `scrapingOffsets` and `hiscoreEntries` optional. These fields only exist when the player is being tracked.